Added block breaking/placing logic. Made config more logical. Added config option for die on disconnect

+++ src/main/java/eagleseye/incombat/logic/PlayerDisconnectLogic.java
@@ -7,7 +7,7 @@ import net.minecraft.server.network.ServerPlayerEntity;
 
 public class PlayerDisconnectLogic {
     private static void killPlayerOnDisconnect(ServerPlayerEntity player){
-        if(CombatCheck.isPlayerInCombat(player)){
+        if(CombatCheck.isPlayerInCombat(player) && InCombat.COMBAT_CONFIG.dieOnDisconnect()){
             player.kill();
         }
     }
